broke out utils into a seperate function to fix strapi detection
